refactor right menu column types and colors.
