+++ src/modules/BurgerMenu/RightMenu/sectionTypes.js
@@ -1,0 +1,11 @@
+// right menu section column/menu types;
+
+export const COLUMN = {
+  POPULAR_SPORTS: 'popular-sports',
+  ALL_SPORTS: 'all-sports',
+  ABOUT_US: 'about-us',
+};
+
+export const MENU = {
+  SPORTS: 'sports',
+};
